back/requirements.txt
asgiref==3.3.1
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
Django==3.1.5
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-swagger==2.2.0
django-searchable-select==1.5.0
djangorestframework==3.12.2
drf-yasg==1.20.0
fonttools==4.26.2
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
Pillow==9.0.0
psycopg2-binary==2.8.6
pycparser==2.20
pydyf==0.0.3
pyparsing==2.4.7
pyphen==0.11.0
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.1.8
celery==4.4.2
django-redis==5.0.0
redis==3.5.3 
pdfkit==0.6.1
xlrd==2.0.1
django-import-export